+++ app/Http/Controllers/RoleType/RoleTypeController.php
@@ -120,6 +120,14 @@ class RoleTypeController extends Controller
      */
     public function destroy($id)
     {
+        $model = $this->repository->find($id);
+        if (!$model) {
+            return responseJson(404, __('message.data not found'));
+        }
+
+        if ($model->hasChildren()) {
+            return responseJson(400,__("this item has children and can't be deleted remove it's children first"));
+        }
         $this->repository->delete($id);
         return responseJson(200, __('deleted'));
     }
